refactor: Removed check active editor from model (#443)
