+++ src/model.ts
@@ -49,6 +49,10 @@ export class Model implements IDisposable {
   public readonly onDidChangeRepository: Event<IModelChangeEvent> = this
     ._onDidChangeRepository.event;
 
+  private _onDidChangeStatusRepository = new EventEmitter<Repository>();
+  public readonly onDidChangeStatusRepository: Event<Repository> = this
+    ._onDidChangeStatusRepository.event;
+
   public openRepositories: IOpenRepository[] = [];
   private disposables: Disposable[] = [];
   private enabled = false;
@@ -139,12 +143,6 @@ export class Model implements IDisposable {
       this.disposables
     );
 
-    window.onDidChangeActiveTextEditor(
-      () => this.checkHasChangesOnActiveEditor(),
-      this,
-      this.disposables
-    );
-
     await this.scanWorkspaceFolders();
   }
 
@@ -178,54 +176,6 @@ export class Model implements IDisposable {
     repository.statusExternal
       .map(r => path.join(repository.workspaceRoot, r.path))
       .forEach(p => this.eventuallyScanPossibleSvnRepository(p));
-  }
-
-  private hasChangesOnActiveEditor(): boolean {
-    if (!window.activeTextEditor) {
-      return false;
-    }
-    const uri = window.activeTextEditor.document.uri;
-
-    const repository = this.getRepository(uri);
-    if (!repository) {
-      return false;
-    }
-
-    const resource = repository.getResourceFromFile(uri);
-    if (!resource) {
-      return false;
-    }
-
-    switch (resource.type) {
-      case Status.ADDED:
-      case Status.DELETED:
-      case Status.EXTERNAL:
-      case Status.IGNORED:
-      case Status.NONE:
-      case Status.NORMAL:
-      case Status.UNVERSIONED:
-        return false;
-      case Status.CONFLICTED:
-      case Status.INCOMPLETE:
-      case Status.MERGED:
-      case Status.MISSING:
-      case Status.MODIFIED:
-      case Status.OBSTRUCTED:
-      case Status.REPLACED:
-        return true;
-    }
-
-    // Show if not match
-    return true;
-  }
-
-  @debounce(100)
-  private checkHasChangesOnActiveEditor() {
-    commands.executeCommand(
-      "setContext",
-      "svnActiveEditorHasChanges",
-      this.hasChangesOnActiveEditor()
-    );
   }
 
   private disable(): void {
@@ -456,15 +406,19 @@ export class Model implements IDisposable {
       this._onDidChangeRepository.fire({ repository, uri })
     );
 
+    const changeStatus = repository.onDidChangeStatus(() => {
+      this._onDidChangeStatusRepository.fire(repository);
+    });
+
     const statusListener = repository.onDidChangeStatus(() => {
       this.scanExternals(repository);
-      this.checkHasChangesOnActiveEditor();
     });
     this.scanExternals(repository);
 
     const dispose = () => {
       disappearListener.dispose();
       changeListener.dispose();
+      changeStatus.dispose();
       statusListener.dispose();
       repository.dispose();
